Add code for testing work done on class creation
The most the of work was done on reorganize class code 
but text cells were created with indications for working 
from remote and from local. It seems that changes on 
class code are not refreshed on notebook working from 
remote. Work on local will be done to test that thought